refactor: types 선언 수정
