+++ yongjae/projects/FEDC5-3_VanillaJS_1/src/components/App.ts
@@ -1,4 +1,5 @@
-import { TodoComponentStatefulProps, Todos } from "../types";
+import { TodoComponentStatefulProps } from "../types/components";
+import { Todos } from "../types/states";
 import { setItem } from "../utils/storage";
 import { validateState } from "../utils/validateState";
 import Header from "./Header";
