+++ yongjae/projects/FEDC5-3_VanillaJS_1/src/components/Header.ts
@@ -1,7 +1,7 @@
 import {
   TodoComponentStatelessContext,
   TodoComponentStatelessProps,
-} from "../types";
+} from "../types/components";
 
 type HeaderProps = TodoComponentStatelessProps<{ text: string }>;
