+++ yongjae/projects/FEDC5-3_VanillaJS_1/src/types/states.ts
@@ -1,0 +1,6 @@
+export interface Todo {
+  text: string;
+  id: string;
+  isCompleted: boolean;
+}
+export type Todos = Todo[];
